Move long tests in test_ham_senzero
